Crean up intra-crate dependencies between cache, housekeeper and unsafe_weak_pointer modules
