+++ src/common.rs
@@ -8,6 +8,7 @@ pub(crate) mod deques;
 pub(crate) mod frequency_sketch;
 pub(crate) mod housekeeper;
 pub(crate) mod thread_pool;
+pub(crate) mod unsafe_weak_pointer;
 
 pub(crate) struct KeyHash<K> {
     pub(crate) key: Arc<K>,
